Move 'LocalMediaItem' to its own file
